+++ tasks.py
@@ -1,66 +1,117 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals, absolute_import
 
-from os.path import join, abspath, dirname
+import os
 
-from invoke import run, task
+from invoke import task, call
 
-ROOT = abspath(join(dirname(__file__)))
+ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__)))
 
-
-def green(text):
-    return '\033[1;32m{0}\033[0;m'.format(text)
+TO_CLEAN = ['build', 'dist', '**/*.pyc', 'reports']
 
 
-def red(text):
-    return '\033[1;31m{0}\033[0;m'.format(text)
+def color(code):
+    '''A simple ANSI color wrapper factory'''
+    return lambda t: '\033[{0}{1}\033[0;m'.format(code, t)
 
 
-def cyan(text):
-    return '\033[1;36m{0}\033[0;m'.format(text)
+green = color('1;32m')
+red = color('1;31m')
+blue = color('1;30m')
+cyan = color('1;36m')
+purple = color('1;35m')
+white = color('1;39m')
 
 
-def lrun(command, *args, **kwargs):
-    run('cd {0} && {1}'.format(ROOT, command), *args, **kwargs)
+def header(text):
+    '''Display an header'''
+    print(' '.join((blue('>>'), cyan(text))))
+
+
+def info(text, *args, **kwargs):
+    '''Display informations'''
+    text = text.format(*args, **kwargs)
+    print(' '.join((purple('>>>'), text)))
+
+
+def success(text):
+    '''Display a success message'''
+    print(' '.join((green('>>'), white(text))))
+
+
+def error(text):
+    '''Display an error message'''
+    print(red('✘ {0}'.format(text)))
 
 
 @task
-def clean(ctx, bower=False, node=False):
+def clean(ctx):
     '''Cleanup all build artifacts'''
-    patterns = [
-        'build', 'dist', 'cover', 'docs/_build',
-        '**/*.pyc', '*.egg-info', '.tox'
-    ]
-    for pattern in patterns:
-        print('Removing {0}'.format(pattern))
-        lrun('rm -rf {0}'.format(pattern))
+    header(clean.__doc__)
+    with ctx.cd(ROOT):
+        for pattern in TO_CLEAN:
+            info('Removing {0}', pattern)
+            ctx.run('rm -rf {0}'.format(pattern))
 
 
 @task
-def test(ctx):
+def test(ctx, report=False):
     '''Run tests suite'''
-    lrun('nosetests --rednose --force-color udata_croquemort', pty=True)
+    cmd = 'pytest -v'
+    if report:
+        cmd = ' '.join((cmd, '--junitxml=reports/tests.xml'))
+    with ctx.cd(ROOT):
+        ctx.run(cmd, pty=True)
 
 
 @task
-def cover(ctx):
+def cover(ctx, html=False):
     '''Run tests suite with coverage'''
-    lrun('nosetests --rednose --force-color --with-coverage '
-         '--cover-html --cover-package=udata_croquemort', pty=True)
+    cmd = 'pytest --cov udata_croquemort --cov-report term'
+    if html:
+        cmd = ' '.join((cmd, '--cov-report html:reports/cover'))
+    with ctx.cd(ROOT):
+        ctx.run(cmd, pty=True)
 
 
 @task
 def qa(ctx):
     '''Run a quality report'''
-    lrun('flake8 udata_croquemort')
+    header(qa.__doc__)
+    with ctx.cd(ROOT):
+        info('Python Static Analysis')
+        flake8_results = ctx.run('flake8 udata_croquemort', pty=True, warn=True)
+        if flake8_results.failed:
+            error('There is some lints to fix')
+        else:
+            success('No lint to fix')
+        info('Ensure PyPI can render README and CHANGELOG')
+        readme_results = ctx.run('python setup.py check -r -s', pty=True, warn=True, hide=True)
+        if readme_results.failed:
+            print(readme_results.stdout)
+            error('README and/or CHANGELOG is not renderable by PyPI')
+        else:
+            success('README and CHANGELOG are renderable by PyPI')
+    if flake8_results.failed or readme_results.failed:
+        error('Quality check failed')
+        exit(flake8_results.return_code or readme_results.return_code)
+    success('Quality check OK')
 
 
-@task(clean)
-def pydist(ctx, buildno=None):
+@task
+def dist(ctx, buildno=None):
     '''Package for distribution'''
-    print(cyan('Building a distribuable package'))
+    header('Building a distribuable package')
     cmd = ['python setup.py']
     if buildno:
         cmd.append('egg_info -b {0}'.format(buildno))
     cmd.append('bdist_wheel')
-    lrun(' '.join(cmd), pty=True)
+    with ctx.cd(ROOT):
+        ctx.run(' '.join(cmd), pty=True)
+    success('Distribution is available in dist directory')
+
+
+@task(clean, qa, call(test, report=True), dist, default=True)
+def default(ctx):
+    '''Perform quality report, tests and packaging'''
+    pass